requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
SQLAlchemy==2.0.23
python-dotenv==1.0.0
schedule==1.2.0